+++ frontend/src/views/Graph.vue
